uint8_t ReadUInt8(int32_t bank_id, size_t addr); 		uint16_t ReadUInt16(int32_t bank_id, size_t addr); 		uint32_t ReadUInt32(int32_t bank_id, size_t addr); 		uint64_t ReadUInt64(int32_t bank_id, size_t addr); 		void ReadBuffer(int32_t bank_id, size_t addr, uint8_t *buffer, int32_t count);

+++ atk-stm32h743-bsp/include/bsp/bsp.cpp
@@ -206,9 +206,9 @@ void TestFlash()
 					flash.EraseSector(2, i, 1);
 				}
 
-				uint32_t value = flash.ReadUInt8(2, 0);
+				uint32_t value = flash.ReadUInt32(2, 0);
 				flash.Program(2, 0, buffer);
-				value = flash.ReadUInt8(2, 0);
+				value = flash.ReadUInt32(2, 0);
 				BSP::GreenDigitalLed().Toggle();
 			}
 		}
